Refactor copy-on-launch-modal into its own component
